A type of the 'id' field from the 'Color' class has been replaced by 'Long'

+++ src/main/java/dao/impl/mysql/CarDaoImpl.java
@@ -65,7 +65,7 @@ class CarDaoImpl extends AbstractDao<Car> implements CarDao{
 
         ps = con.prepareStatement(sqlInsert, Statement.RETURN_GENERATED_KEYS);
         ps.setString(1, entity.getNumber());
-        ps.setInt(2, entity.getColor().getId());
+        ps.setLong(2, entity.getColor().getId());
         ps.setInt(3, entity.getModel().getId());
 
         return ps;
